+++ src/oms_sensemaking/config.py
@@ -116,7 +116,10 @@ class MilSymbolSettings(BaseModel):
     attribute_code_iris: list[str] = Field(
         ["http://www.ontologyrepository.com/CommonCoreOntologies/has_text_value"],
         description="Attribute Iris for full mil symbol codes")
-
+    b_c_placeholders: list[str] = Field(["*", "-"],
+                                description="Placeholder values for unset values in 2525B and 2525C symbol Id codes")
+    b_c_standard_placeholder: str = Field("-",
+                    description="Standardized placeholder string for unset values in 2525B and 2525C symbol Id codes")
     # War, Pending, Unknown, Present
     default_2525b_code: str = Field(
         "SUZP-----------", description="Default 2525B code")